feat: improve styles
