+++ src/components/AboutSection/About/About.module.scss
@@ -1,19 +1,31 @@
 @use '@/styles/mixins.scss' as mix;
 @use '@/styles/variables.scss' as vars;
 
+.section {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+  justify-content: center;
+  margin: 0 auto;
+
+  @media (width <= 768px) {
+    flex-direction: column;
+    align-items: center;
+  }
+}
+
 .about {
   position: relative;
   display: flex;
   flex-direction: column;
-  gap: 3rem;
-  min-height: 100vh;
-  padding: 3rem 1.5rem;
-  padding-bottom: 8rem;
+  gap: 2rem;
+  min-height: 100%;
+  padding-bottom: 15rem;
 
   h1 {
-    @include mix.headline-1(vars.$color-black, 2rem);
-
     text-align: center;
+
+    @include mix.headline-1(vars.$color-black, 2rem);
   }
 }
 
@@ -22,8 +34,8 @@
   z-index: 10;
   right: 2rem;
   bottom: 2rem;
-  width: 300px;
-  opacity: 0.8;
+  width: 13rem;
+  filter: opacity(0.5);
   transition: all 0.3s ease-in-out;
 
   img {
@@ -32,30 +44,29 @@
     height: auto;
   }
 
-  &:hover {
-    transform: translateY(-3px);
-    opacity: 1;
+  @media (hover: hover) {
+    &:hover {
+      cursor: pointer;
+      scale: 1.05;
+      filter: opacity(1);
+    }
   }
 }
 
-@media (width <= 1100px) and (width >= 769px) {
+@media (width <= 1050px) {
   .logo {
     right: 1rem;
     bottom: 1rem;
-    width: 200px;
+    width: 10rem;
   }
 
   .about {
-    padding-bottom: 6rem;
+    padding-bottom: 1rem;
   }
 }
 
 @media (width <= 768px) {
   .about {
-    gap: 2rem;
-    padding: 2rem 1rem;
-    padding-bottom: 5rem;
-
     h1 {
       font-size: 2rem;
     }
@@ -66,7 +77,7 @@
     right: auto;
     bottom: auto;
     align-self: center;
-    width: 150px;
+    width: 8rem;
     margin-top: 1rem;
   }
 }
