+++ src/components/AboutSection/AboutSection.tsx
@@ -1,42 +1,76 @@
 import type { ReactElement } from 'react'
+import { useEffect, useState } from 'react'
+import { About } from './About/About'
 import s from './AboutSection.module.scss'
-import AboutCard, { type AboutCardProps } from '@/components/AboutCard/AboutCard'
+import { Button } from '@/shared/ui/Button/Button'
+import { RoutePath } from '@/shared/config/routeConfig/routeConfig'
+import { Link } from 'react-router-dom'
 
-const ArtemDataProps: AboutCardProps = {
-  name: 'Artem Gassan',
-  activities: ['feature', 'styles', 'refactor'],
-  gitHubName: 'artemgassan',
-  gitHubLink: 'https://github.com/artemgassan',
-  avatarLink: 'https://avatars.githubusercontent.com/u/158855420?v=4',
-  description: 'There will be a description here later',
-}
+export const AboutSection = (): ReactElement => {
+  const [magicOne, setMagicOne] = useState(true)
+  const [magicTwo, setMagicTwo] = useState(false)
+  const [magicThree, setMagicThree] = useState(false)
+  const [about, setAbout] = useState(false)
+  const TIME_OF_MAGIC = 2000
+  useEffect(() => {
+    if (magicThree) {
+      const timer = setTimeout(() => {
+        setAbout(true)
+        setMagicThree(false)
+      }, TIME_OF_MAGIC)
 
-const DaryaDataProps: AboutCardProps = {
-  name: 'Darya Kolenchenko',
-  activities: ['tests', 'CI/CD', 'API'],
-  gitHubName: 'dariechka',
-  gitHubLink: 'https://github.com/dariechka',
-  avatarLink: 'https://avatars.githubusercontent.com/u/149780473?v=4',
-  description: 'There will be a description here later',
-}
-
-const AnnaDataProps: AboutCardProps = {
-  name: 'Anna Vasilevich',
-  activities: ['routing', 'feature', 'chore'],
-  gitHubName: 'dzichonka',
-  gitHubLink: 'https://github.com/dzichonka',
-  avatarLink: 'https://avatars.githubusercontent.com/u/73832561?v=4',
-  description: 'There will be a description here later',
-}
-
-const AboutSection = (): ReactElement => {
+      return (): void => clearTimeout(timer)
+    }
+  }, [magicThree])
   return (
-    <section className={s.section}>
-      <AboutCard {...DaryaDataProps} />
-      <AboutCard {...AnnaDataProps} />
-      <AboutCard {...ArtemDataProps} />
+    <section className={`section ${s.section} ${magicThree ? s.white : ''}`}>
+      {magicOne && (
+        <div className={s.magic}>
+          <p>To uncover the secret knowledge of the Sorcerer's Apprentices, you must cast a powerful spell...</p>
+          <p>
+            Say the magic words <span className={s.abracadabra}>"Abracadabra!"</span> and then click the button below.
+          </p>
+          <Button
+            onClick={() => {
+              setMagicTwo(true)
+              setMagicOne(false)
+            }}
+          >
+            Abracadabra!
+          </Button>
+        </div>
+      )}
+      {magicTwo && (
+        <div className={s.magic}>
+          <p>No, no, no... </p>
+          <p>
+            You must say <span className={s.abracadabra}>"Abracadabra!"</span> loudly and clearly!
+          </p>
+          <p>Or perhaps... you're just a Muggle?</p>
+          <div className={s.buttons}>
+            <Button
+              onClick={() => {
+                setMagicThree(true)
+                setMagicTwo(false)
+              }}
+            >
+              ABRAKADABRA!
+            </Button>
+            <Link to={RoutePath.MAIN}>
+              <Button onClick={() => setMagicTwo(false)}>I'm a Muggle :(</Button>
+            </Link>
+          </div>
+        </div>
+      )}
+      {magicThree && (
+        <div className={s.loading}>
+          <p>
+            Secret knowledge <br />
+            is loading...
+          </p>
+        </div>
+      )}
+      {about && <About />}
     </section>
   )
 }
-
-export default AboutSection
