+++ src/components/HeroSection/HeroSection.module.scss
@@ -10,7 +10,7 @@
   justify-content: flex-end;
   height: 30rem;
   padding: 5rem;
-  background-color: vars.$color-primary200;
+  background-color: vars.$color-primary300;
   background-image: url('/images/cat-transparent-bg.png');
   background-repeat: no-repeat;
   background-position: bottom right;
