--- src/components/Magic/Magic.module.scss
@@ -1,17 +1,0 @@
-@use '@/styles/variables.scss' as vars;
-@use '@/styles/mixins.scss' as mixins;
-
-.magic {
-  min-height: 40vh;
-
-  @include mixins.center(1rem);
-}
-
-.abracadabra {
-  font-weight: 600;
-  color: vars.$state-color-error;
-}
-
-.buttons {
-  @include mixins.flex-row;
-}
