+++ src/pages/AboutPage/AboutPage.tsx
@@ -1,14 +1,14 @@
 import s from './AboutPage.module.scss'
 import type { ReactElement } from 'react'
 import Footer from '@/components/Footer/Footer'
-import { Magic } from '@/components/Magic/Magic'
+import { AboutSection } from '@/components/AboutSection/AboutSection'
 import { Header } from '@/components/Header/Header'
 
 const AboutPage = (): ReactElement => {
   return (
     <div className={s.page}>
       <Header />
-      <Magic />
+      <AboutSection />
       <Footer />
     </div>
   )
